Written also Mapper and reducer, hope that this work

+++ src/main/java/it/unibo/bd1819/reducers/SortReducer.java
@@ -9,8 +9,8 @@ import java.io.IOException;
 
 import static it.unibo.bd1819.utils.Separators.CUSTOM_VALUE_SEPARATOR;
 
-public class SortReducer extends Reducer<LongWritable, Text, Text, Text> {
-    public void reduce(LongWritable key, Iterable<Text> values,
+public class SortReducer extends Reducer<IntWritable, Text, Text, Text> {
+    public void reduce(IntWritable key, Iterable<Text> values,
                        Context context
     ) throws IOException, InterruptedException {
         for(Text value: values) {
@@ -22,7 +22,7 @@ public class SortReducer extends Reducer<LongWritable, Text, Text, Text> {
         return new Text("Director:" + oldValue.toString().split(CUSTOM_VALUE_SEPARATOR)[0]);
     }
 
-    private Text extractNewValue(final Text oldValue, final LongWritable moviesDirected) {
+    private Text extractNewValue(final Text oldValue, final IntWritable moviesDirected) {
         final String directorID = oldValue.toString().split(CUSTOM_VALUE_SEPARATOR)[0] + CUSTOM_VALUE_SEPARATOR;
         return new Text("Movies Directed: " + moviesDirected.get() + ", Most Frequently actors: " +
                 oldValue.toString().replace(directorID,""));
